+++ Domain/RDD.Domain.Tests/DefaultFixture.cs
@@ -1,4 +1,8 @@
-﻿using Rdd.Domain.Mocks;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.DependencyInjection.Extensions;
+using Rdd.Application;
+using Rdd.Domain.Helpers.Reflection;
+using Rdd.Domain.Mocks;
 using Rdd.Domain.Models;
 using Rdd.Domain.Patchers;
 using Rdd.Domain.Rights;
@@ -10,16 +14,29 @@ namespace Rdd.Domain.Tests
 {
     public class DefaultFixture : IDisposable
     {
+        public IServiceProvider ServiceProvider { get; private set; }
         public IRightExpressionsHelper<User> RightsService { get; private set; }
-        public IPatcherProvider PatcherProvider { get; private set; }
+        public IPatcherProvider PatcherProvider => ServiceProvider.GetService<IPatcherProvider>();
+        public IReflectionHelper ReflectionHelper => ServiceProvider.GetService<IReflectionHelper>();
         public IInstanciator<User> Instanciator { get; private set; }
         public InMemoryStorageService InMemoryStorage { get; private set; }
         public IRepository<User> UsersRepo { get; private set; }
 
         public DefaultFixture()
         {
+            var services = new ServiceCollection();
+
+            services.TryAddSingleton<IReflectionHelper, ReflectionHelper>();
+            services.TryAddSingleton<IPatcherProvider, PatcherProvider>();
+            services.TryAddSingleton<EnumerablePatcher>();
+            services.TryAddSingleton<DictionaryPatcher>();
+            services.TryAddSingleton<ValuePatcher>();
+            services.TryAddSingleton<DynamicPatcher>();
+            services.TryAddSingleton<ObjectPatcher>();
+
+            ServiceProvider = services.BuildServiceProvider();
+
             RightsService = new RightsServiceMock<User>();
-            PatcherProvider = new PatcherProvider();
             Instanciator = new DefaultInstanciator<User>();
             InMemoryStorage = new InMemoryStorageService();
             UsersRepo = new Repository<User>(InMemoryStorage, RightsService);