+++ Web/RDD.Web/Serialization/Serializers/BaseClassSerializer.cs
@@ -1,17 +1,16 @@
 ﻿using Newtonsoft.Json.Serialization;
 using Rdd.Domain.Helpers.Expressions;
+using Rdd.Domain.Helpers.Reflection;
 using Rdd.Web.Querying;
 using Rdd.Web.Serialization.Providers;
-using Rdd.Web.Serialization.Reflection;
 using Rdd.Web.Serialization.UrlProviders;
-using System;
 
 namespace Rdd.Web.Serialization.Serializers
 {
     public class BaseClassSerializer : EntitySerializer
     {
-        public BaseClassSerializer(ISerializerProvider serializerProvider, IReflectionProvider reflectionProvider, NamingStrategy namingStrategy, IUrlProvider urlProvider, Type workingType)
-            : base(serializerProvider, reflectionProvider, namingStrategy, urlProvider, workingType) { }
+        public BaseClassSerializer(ISerializerProvider serializerProvider, IReflectionHelper reflectionHelper, NamingStrategy namingStrategy, IUrlProvider urlProvider)
+            : base(serializerProvider, reflectionHelper, namingStrategy, urlProvider) { }
 
         protected override IExpressionTree CorrectFields(object entity, IExpressionTree fields)
         {