+++ Web/RDD.Web/Serialization/Serializers/CultureInfoSerializer.cs
@@ -1,9 +1,8 @@
 ﻿using Newtonsoft.Json;
 using Newtonsoft.Json.Serialization;
-using Rdd.Domain.Helpers;
 using Rdd.Domain.Helpers.Expressions;
+using Rdd.Domain.Helpers.Reflection;
 using Rdd.Web.Serialization.Providers;
-using Rdd.Web.Serialization.Reflection;
 using System.Collections.Generic;
 using System.Reflection;
 
@@ -13,8 +12,8 @@ namespace Rdd.Web.Serialization.Serializers
     {
         private static HashSet<string> _allowedProperties = new HashSet<string> { "id", "name", "code" };
 
-        public CultureInfoSerializer(ISerializerProvider serializerProvider, IReflectionProvider reflectionProvider, NamingStrategy namingStrategy)
-            : base(serializerProvider, reflectionProvider, namingStrategy, typeof(Culture)) { }
+        public CultureInfoSerializer(ISerializerProvider serializerProvider, IReflectionHelper reflectionHelper, NamingStrategy namingStrategy)
+            : base(serializerProvider, reflectionHelper, namingStrategy) { }
 
         protected override void SerializeProperty(JsonTextWriter writer, object entity, IExpressionTree fields, PropertyInfo property)
         {