+++ Web/RDD.Web/Serialization/Serializers/DictionarySerializer.cs
@@ -2,8 +2,8 @@ using Newtonsoft.Json;
 using Newtonsoft.Json.Serialization;
 using Rdd.Domain.Exceptions;
 using Rdd.Domain.Helpers.Expressions;
+using Rdd.Domain.Helpers.Reflection;
 using Rdd.Web.Serialization.Providers;
-using Rdd.Web.Serialization.Reflection;
 using System.Collections;
 using System.Linq;
 
@@ -11,8 +11,8 @@ namespace Rdd.Web.Serialization.Serializers
 {
     public class DictionarySerializer : ObjectSerializer
     {
-        public DictionarySerializer(ISerializerProvider serializerProvider, IReflectionProvider reflectionProvider, NamingStrategy namingStrategy) 
-            : base(serializerProvider, reflectionProvider, namingStrategy, typeof(IDictionary)) { }
+        public DictionarySerializer(ISerializerProvider serializerProvider, IReflectionHelper reflectionHelper, NamingStrategy namingStrategy) 
+            : base(serializerProvider, reflectionHelper, namingStrategy) { }
 
         public override void WriteJson(JsonTextWriter writer, object entity, IExpressionTree fields)
             => WriteJson(writer, entity as IDictionary, fields);