+++ Web/RDD.Web/Serialization/Serializers/ObjectSerializer.cs
@@ -1,8 +1,8 @@
 ﻿using Newtonsoft.Json;
 using Newtonsoft.Json.Serialization;
 using Rdd.Domain.Helpers.Expressions;
+using Rdd.Domain.Helpers.Reflection;
 using Rdd.Web.Serialization.Providers;
-using Rdd.Web.Serialization.Reflection;
 using System;
 using System.Linq;
 using System.Linq.Expressions;
@@ -13,17 +13,14 @@ namespace Rdd.Web.Serialization.Serializers
     public class ObjectSerializer : ISerializer
     {
         protected ISerializerProvider SerializerProvider { get; private set; }
-        protected IReflectionProvider ReflectionProvider { get; private set; }
+        protected IReflectionHelper ReflectionHelper { get; private set; }
         protected NamingStrategy NamingStrategy { get; private set; }
 
-        public Type WorkingType { get; set; }
-
-        public ObjectSerializer(ISerializerProvider serializerProvider, IReflectionProvider reflectionProvider, NamingStrategy namingStrategy, Type workingType)
+        public ObjectSerializer(ISerializerProvider serializerProvider, IReflectionHelper reflectionHelper, NamingStrategy namingStrategy)
         {
-            SerializerProvider = serializerProvider;
-            NamingStrategy = namingStrategy;
-            ReflectionProvider = reflectionProvider;
-            WorkingType = workingType;
+            SerializerProvider = serializerProvider ?? throw new ArgumentNullException(nameof(serializerProvider));
+            ReflectionHelper = reflectionHelper ?? throw new ArgumentNullException(nameof(reflectionHelper));
+            NamingStrategy = namingStrategy ?? throw new ArgumentNullException(nameof(namingStrategy));
         }
 
         public virtual void WriteJson(JsonTextWriter writer, object entity, IExpressionTree fields)
@@ -42,8 +39,8 @@ namespace Rdd.Web.Serialization.Serializers
         {
             if (fields == null || !fields.Children.Any())
             {
-                return new ExpressionParser().ParseTree(WorkingType,
-                    string.Join(",", ReflectionProvider.GetProperties(WorkingType).Select(p => p.Name)));
+                var type = entity.GetType();
+                return new ExpressionParser().ParseTree(entity.GetType(), string.Join(",", ReflectionHelper.GetProperties(type).Select(p => p.Name)));
             }
 
             return fields;
@@ -73,6 +70,6 @@ namespace Rdd.Web.Serialization.Serializers
             => NamingStrategy.GetPropertyName(property.Name, false);
 
         protected virtual object GetRawValue(object entity, IExpressionTree fields, PropertyInfo property)
-            => property.GetValue(entity, null);
+            => ReflectionHelper.GetValue(entity, property);
     }
 }